Optionally preserve unknown object fields when both decoding and converting.
This allows the original object being converted/decoded to be re-converted/encoded to an identical length buffer.
The buffer may not necessarily be the identical to the original as fields may be re-serialized in a different order.

This property is useful for proxies/brokers/gateways that would like to inspect a message, potentially make changes
to a subset of the fields, and then pass the message along.  It allows such proxies to only be aware of the fields
they are interested in.